requirements.txt
streamlit>=1.30.0
anthropic>=0.18.0
google-generativeai>=0.4.0
PyPDF2>=3.0.0
python-docx>=1.0.0
pandas>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
